+++ JBQQuizMe.ViewModel/PreferenceKeys.cs
@@ -1,0 +1,13 @@
+﻿using System;
+
+namespace JBQQuizMe.ViewModel
+{
+    public static class PreferenceKeys
+    {
+        public const string ReadQuestions = "read_questions";
+
+        public const string StartQuestionRange = "start_question_range";
+
+        public const string EndQuestionRange = "end_question_range";
+    }
+}