+++ JBQQuizMeApp/QuizPage.xaml.cs
@@ -2,7 +2,7 @@
 
 public partial class QuizPage : ContentPage
 {
-	public QuizPage(int? startQuestionNumber, int? endQuestionNumber)
+	public QuizPage(int? startQuestionNumber, int? endQuestionNumber, bool readQuestions)
 	{
         InitializeComponent();
 
@@ -10,8 +10,11 @@ public partial class QuizPage : ContentPage
 
         context.StartQuestionNumber = startQuestionNumber;
         context.EndQuestionNumber = endQuestionNumber;
+        context.ReadQuestions = readQuestions;
 
+#pragma warning disable CS4014 // Because this call is not awaited, execution of the current method continues before the call is completed
         context.Initialize();
+#pragma warning restore CS4014 // Because this call is not awaited, execution of the current method continues before the call is completed
     }
 }