+++ CHANGELOG.md
@@ -43,3 +43,10 @@ Track changes made to the game application.
 - Display how many candles have been lit
 - Show annoying frog on wrong answers
 - Animate congratulation message when candle is fully lit
+
+## Version 0.6.0.0
+
+- Fourth beta version
+- Add text to speech for questions and answers
+- Save text to speech settings between game runs
+- Save question start and end range between game runs